backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
pymupdf==1.24.10
python-docx==1.1.2
sentence-transformers==3.1.0
faiss-cpu==1.8.0
scikit-learn==1.5.2
numpy==1.26.4
asyncio-timeout==4.0.3
